IBX-11914: Moved file size formatter to helper
