+++ src/bundle/Resources/public/js/scripts/fieldType/ibexa_media.js
@@ -1,3 +1,5 @@
+import { formatBytesToLargestUnit } from '../helpers/bytes.helper';
+
 (function (global, doc, ibexa) {
     const SELECTOR_FIELD = '.ibexa-field-edit--ibexa_media';
     const SELECTOR_PREVIEW = '.ibexa-field-edit__preview';
@@ -52,7 +54,7 @@
             const nameContainer = preview.querySelector('.ibexa-field-edit-preview__file-name');
             const sizeContainer = preview.querySelector('.ibexa-field-edit-preview__file-size');
             const files = [].slice.call(event.target.files);
-            const fileSize = this.formatFileSize(files[0].size);
+            const fileSize = formatBytesToLargestUnit(files[0].size);
             const { escapeHTML } = ibexa.helpers.text;
             const fileName = escapeHTML(files[0].name);
